Fix paths and download process for streamlit cloud
Files can only be saved temporarily on streamlit community cloud.
Temporary folders are created with the application as base folder.
Added Zip creation and zip download for multiple extraction files.
Streamlit can only download a single file with a download button.
This needs to be adapted when output will be in a single file.